pyproject.toml
[project]
name = "unifi-network-mcp"
version = "0.1.0b1"
description = "Unifi Network MCP Server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.6.0",
    "aiohttp>=3.8.5",
    "aiounifi>=83.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "omegaconf>=2.3.0",
    "jsonschema>=4.17.0",
    "typing-extensions>=4.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
# CLI entrypoint that launches the UniFi Network MCP server
mcp-server-unifi-network = "src.main:main"
